+++ src/protocols/media/sdmmc.rs
@@ -1,0 +1,28 @@
+/*
+ * Dyspxkrypt LibUEFI: Raw bindings of UEFI that conforms to the definitions of the UEFI Specification.
+ * Copyright (C) 2023-2024 HTGAzureX1212.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+use crate::types::EFI_GUID;
+
+pub const EFI_SD_MMC_PASS_THRU_PROTOCOL_GUID: EFI_GUID = unsafe {
+    EFI_GUID::from_raw_parts(
+        0x716EF0D9,
+        0xFF83,
+        0x4F69,
+        [0x81, 0xE9, 0x51, 0x8B, 0xD3, 0x9A, 0x8E, 0x70],
+    )
+};