+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## Added
 
-- EFI NVME pass thru protocol and its associating constants and structures (`EFI_NVM_EXPRESS_PASS_THRU_PROTOCOL`) (behind `nvme` or `media-full` feature flags)
-- EFI SD MMC pass thru protocol and its associating constants and structures (`EFI_SD_MMC_PASS_THRU_PROTOCOL`) (behind `sd-mmc` or `media-full` feature flags)
+- EFI NVME pass thru protocol and its associating constants and structures (`EFI_NVM_EXPRESS_PASS_THRU_PROTOCOL`) (behind `media-nvme` or `media-full` feature flags)
+- EFI SD MMC pass thru protocol and its associating constants and structures (`EFI_SD_MMC_PASS_THRU_PROTOCOL`) (behind `media-sd-mmc` or `media-full` feature flags)
+- EFI ramdisk protocol and its associating constants and structures (`EFI_RAM_DISK_PROTOCOL`) (behind `media-ramdisk` or `media-full` feature flags)
 
 ## Changed