+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - EFI SD MMC pass thru protocol and its associating constants and structures (`EFI_SD_MMC_PASS_THRU_PROTOCOL`) (behind `media-sd-mmc` or `media-full` feature flags)
 - EFI ramdisk protocol and its associating constants and structures (`EFI_RAM_DISK_PROTOCOL`) (behind `media-ramdisk` or `media-full` feature flags)
 - EFI partition info protocol and its associating constants and structures (`EFI_PARTITION_INFO_PROTOCOL`) (behind `media-partition` or `media-full` feature flags)
+- EFI NVDIMM protocol and its associating constants and structures (`EFI_NVDIMM_LABEL_PROTOCOL`) (behind `media-nvdimm` or `media-full` feature flags)
 
 ## Changed